requirements.txt
calmap==0.0.9
pandas==1.5.2
python-dotenv==0.21.0
requests==2.28.1
streamlit==1.17.0